docs: document --show-prompt option
