+++ src/index.js
@@ -5,6 +5,7 @@ import "dotenv/config";
 import { Command } from "commander";
 import path from "node:path";
 import { DEFAULT_PROMPT_PATH } from "./config.js";
+import { buildPrompt } from "./prompt.js";
 
 const program = new Command();
 program
@@ -32,10 +33,11 @@ program
     "-x, --context <file>",
     "Text file with exhibition context for the curators"
   )
+  .option("--show-prompt", "Print the final prompt and exit")
   .option("--no-recurse", "Process a single directory only")
   .parse(process.argv);
 
-const { dir, prompt: promptPath, model, recurse, apiKey, curators, context: contextPath } = program.opts();
+const { dir, prompt: promptPath, model, recurse, apiKey, curators, context: contextPath, showPrompt } = program.opts();
 
 if (apiKey) {
   process.env.OPENAI_API_KEY = apiKey;
@@ -51,13 +53,17 @@ if (apiKey) {
     }
     const absDir = path.resolve(dir);
     const { triageDirectory } = await import("./orchestrator.js");
+    const prompt = await buildPrompt(promptPath, { curators, contextPath });
+    if (showPrompt) {
+      console.log(prompt);
+      return;
+    }
     await triageDirectory({
       dir: absDir,
-      promptPath,
+      prompt,
       model,
       recurse,
       curators,
-      contextPath,
     });
     console.log("🎉  Finished triaging.");
   } catch (err) {
